Code quality: type aliases, clippy fixes, doc alignment
- Add ForwardOutput and DetectOutput type aliases for complex return types
- Remove unnecessary `as i32` cast (input_size already i32)
- Remove unnecessary `as f32` casts (image_center already f32)
- Fix partial_cmp borrowing (pass by ref not double-ref)
- Use field init shorthand for fmc
- Align doc comment continuation lines to 3-space indent

+++ src/helpers/opencv_helper.rs
@@ -111,7 +111,7 @@ impl OpenCVHelper {
         let blob = dnn::blob_from_image(
             &image,
             1.0 / self.std as f64,
-            core::Size::new(input_size.0 as i32, input_size.1 as i32),
+            core::Size::new(input_size.0, input_size.1),
             core::Scalar::new(self.mean as f64, self.mean as f64, self.mean as f64, 0.0),
             true,
             false,
